Refactoring: make ProcessHandler an inner class of ProcessPool.
